EnglishForce-AI/AIServer/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
apscheduler==3.10.4
tensorflow==2.19.0
nltk==3.8.1
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
python-multipart==0.0.6 